fix(core): restore tray icon after Tauri upgrade to v2
