+++ src-tauri/Cargo.toml
@@ -13,10 +13,17 @@ authors = ["Ivan Gabriele"]
 homepage = "https://github.com/ivangabriele/clamav-desktop"
 repository = "https://github.com/ivangabriele/clamav-desktop.git"
 
-[features]
-# this feature is used for production builds or when `devPath` points to the filesystem
-# DO NOT REMOVE!!
-custom-protocol = ["tauri/custom-protocol"]
+[lib]
+# The `_lib` suffix may seem redundant but it is necessary
+# to make the lib name unique and wouldn't conflict with the bin name.
+# This seems to be only an issue on Windows, see https://github.com/rust-lang/cargo/issues/8519
+name = "clamav_desktop_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+# [features]
+# # this feature is used for production builds or when `devPath` points to the filesystem
+# # DO NOT REMOVE!!
+# custom-protocol = ["tauri/custom-protocol"]
 
 [workspace]
 members = ["cli", "common", "config", "dev", "fast-cli", "filer"]
