+++ src-tauri/src/system/tray.rs
@@ -1,22 +1,25 @@
 use tauri::{
-    menu::{Menu, MenuItem},
+    menu::{self, Menu, MenuItem},
     tray::{MouseButton, MouseButtonState, TrayIconBuilder, TrayIconEvent},
-    Manager, Runtime,
+    AppHandle, Manager, Runtime,
 };
 
-#[cfg(not(tarpaulin_include))]
-pub fn new<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
-    let quit_i = MenuItem::with_id(app, "quit", "Quit", true, None::<&str>)?;
-    let menu = Menu::with_items(app, &[&quit_i])?;
+use crate::globals;
 
-    let _ = TrayIconBuilder::with_id("tray")
-        .icon(app.default_window_icon().unwrap().clone())
+use super::window;
+
+#[cfg(not(tarpaulin_include))]
+pub fn new_tray_icon<R: Runtime>(app_handle: &AppHandle<R>) -> tauri::Result<()> {
+    let menu = new_menu(app_handle, false);
+
+    let _ = TrayIconBuilder::with_id(globals::MAIN_TRAY_ICON_ID)
+        .icon(app_handle.default_window_icon().unwrap().clone())
         .menu(&menu)
         .menu_on_left_click(false)
-        .on_menu_event(move |app, event| match event.id.as_ref() {
-            "quit" => {
-                app.exit(0);
-            }
+        .on_menu_event(move |app_handle, event| match event.id.as_ref() {
+            "show" => window::show(app_handle),
+            "hide" => window::hide(app_handle),
+            "quit" => app_handle.exit(0),
             _ => {}
         })
         .on_tray_icon_event(|tray, event| {
@@ -26,24 +29,26 @@ pub fn new<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
                 ..
             } = event
             {
-                let app = tray.app_handle();
-                if let Some(window) = app.get_webview_window("main") {
-                    let _ = window.show();
-                    let _ = window.set_focus();
-                }
+                let window = tray
+                    .app_handle()
+                    .get_window(globals::MAIN_TRAY_ICON_ID)
+                    .expect("Could not get window.");
+
+                let _ = window.show();
+                let _ = window.set_focus();
             }
         })
-        .build(app);
+        .build(app_handle);
 
     Ok(())
 }
 
-// #[cfg(not(tarpaulin_include))]
-// pub fn new() -> SystemTray {
-//     let tray_menu = SystemTrayMenu::new()
-//         .add_item(CustomMenuItem::new("toggle".to_string(), "Hide"))
-//         .add_item(CustomMenuItem::new("quit".to_string(), "Quit"));
-//     let system_tray = SystemTray::new().with_menu(tray_menu);
+pub fn new_menu<R: Runtime>(app_handle: &tauri::AppHandle<R>, is_windows_hidden: bool) -> menu::Menu<R> {
+    let toggle_menu_item = match is_windows_hidden {
+        true => MenuItem::with_id(app_handle, "show", "Show", true, None::<&str>).unwrap(),
+        false => MenuItem::with_id(app_handle, "hide", "Hide", true, None::<&str>).unwrap(),
+    };
+    let quit_menu_item = MenuItem::with_id(app_handle, "quit", "Quit", true, None::<&str>).unwrap();
 
-//     system_tray
-// }
+    Menu::with_items(app_handle, &[&toggle_menu_item, &quit_menu_item]).unwrap()
+}
