+++ src/screens/Scanner/index.tsx
@@ -1,4 +1,4 @@
-import { invoke } from '@tauri-apps/api'
+import { invoke } from '@tauri-apps/api/core'
 import { listen } from '@tauri-apps/api/event'
 import numeral from 'numeral'
 import { useCallback, useEffect } from 'react'
